:recycle: refactor: BadRequestException들 경로 리팩터링
